refactor select triggers to new file
