+++ src/commands/actions.js
@@ -1,43 +1,12 @@
-import chalk from 'chalk'
-
 import prompt from '../libs/inquirer'
 import event from '../core/event'
-import filter from '../core/filter'
 import messages from '../messages'
 import actions from '../actions'
 import actionsCore from '../core/actions'
 import ServiceAccounts from '../serviceAccounts'
 import Questions from './questions/action'
 import config from '../core/config'
-import listener from '../core/listener'
-
-const { getTriggerModels, loadTriggers } = listener
-
-const validTriggers = ['onSuccess', 'onFailure']
-
-const checkTrigger = trigger => validTriggers.includes(trigger)
-
-function printsInvalidTriggerError() {
-  messages.error('Inexistent trigger. Please specify a valid trigger.');
-  const availableTriggersInfo = validTriggers.map(trigger => ` ${chalk.blue.bold(`${trigger}`)}`);
-  messages.info(`Available triggers are:${availableTriggersInfo}`);
-}
-
-const checkEvent = (eventId) => {
-  if (!event.eventExists(eventId)) {
-    messages.error(`${chalk.blue.bold(eventId)} does not exist!`)
-    return false
-  }
-  return true
-}
-
-const checkFilter = (eventId, filterId) => {
-  if (!filter.filterExists(eventId, filterId)) {
-    messages.error(`Filter ${chalk.blue.bold(filterId)} does not exist on ${chalk.blue.bold(eventId)}.`)
-    return false
-  }
-  return true
-}
+import checks from './check'
 
 const actionQuestions = async (actionConfig, { id, eventId, filterId, type, action, serviceAccountId, async }) => {
   try {
@@ -90,8 +59,8 @@ const addAction = (program) => {
     .action(async (eventId, filterId, type, action) => {
       try {
         config.getConfigurationFile()
-        if (!checkEvent(eventId)) return false
-        if (!checkFilter(eventId, filterId)) return false
+        if (!checks.checkEvent(eventId)) return false
+        if (!checks.checkFilter(eventId, filterId)) return false
         const actionConfig = actions.getActionConfig(type, action)
         if (!actionConfig) return false
         const serviceAccounts = await ServiceAccounts.getServiceAccountsOfType(type)
@@ -121,47 +90,6 @@ const addAction = (program) => {
     })
 }
 
-const addTrigger = (program) => {
-  program
-    .command('add:trigger <triggerEvent> <eventId> <filterId> <actionId>')
-    .description('Add a new trigger to an action')
-    .action(async (triggerEvent, eventId, filterId, actionId) => {
-      try {
-        config.getConfigurationFile()
-        if (!checkEvent(eventId)) return false
-        if (!checkFilter(eventId, filterId)) return false
-        const actionExists = actionsCore.actionExists(eventId, filterId, actionId)
-        if (!actionExists) return false
-        const isValidTrigger = checkTrigger(triggerEvent)
-        if (!isValidTrigger) {
-          printsInvalidTriggerError()
-          return false
-        }
-
-        const eventsToBeCalled = await loadTriggers()
-        // TODO verificar recursão de eventos -> caso eventsToBeCalled contenha eventId, o que faremos?
-        const triggerModels = await getTriggerModels(eventsToBeCalled)
-
-        const created = actionsCore.createTrigger({ triggerEvent, triggerModels, eventId, filterId, actionId })
-
-        if (created) {
-          const eventIds = eventsToBeCalled.map(e => e.eventId)
-          messages.success(`Trigger created: ${chalk.blue.bold(eventIds)} will be called ` +
-            `under the condition ${chalk.blue.bold(triggerEvent)} ` +
-            `of action action ${chalk.blue.bold(`${actionId}`)} on event ${chalk.blue.bold(`${eventId}:${filterId}`)}.`)
-          // messages.info('Check the action template at the YAML file and replace its ') // TODO add next actions tip pointing to documentation
-        }
-
-        return created
-      } catch (e) {
-        console.error(e)
-        return false
-      }
-    })
-}
-
-
 export default {
   addAction,
-  addTrigger,
 }
